Bill Service, Bill Controller, Bill Repository
'double' to 'decimal' type for 'Amount' property in transactions, dues collections

Enum to string convertor at json in transactions, dues, bills entity

+++ src/BudgetTracker.Core/Domain/Entities/Due.cs
@@ -15,7 +15,7 @@ public class Due : MongoObject
     public string Payee { get; set; } = string.Empty;
 
     [JsonPropertyName("principal_amount"), BsonElement("principal_amount")]
-    public double PrincipalAmount { get; set; }
+    public decimal PrincipalAmount { get; set; }
 
     [Required, JsonPropertyName("description"), BsonElement("description")]
     public string Description { get; set; } = string.Empty;
